modify math_functor.cpp
