+++ oneflow/core/functional/impl/array_functor.cpp
@@ -712,10 +712,10 @@ class SliceUpdateFunctor {
     JUST(attrs.SetAttr<std::vector<int64_t>>("step", step));
     if (inplace) {
       JUST(CheckInplaceValid(x));
+      CHECK_OR_RETURN(JUST(IsContiguous(x)));
       auto outputs = std::make_shared<TensorTuple>(1);
       outputs->at(0) = x;
-      JUST(OpInterpUtil::Dispatch(*op_, {x->contiguous(), update->contiguous()}, outputs.get(),
-                                  attrs));
+      JUST(OpInterpUtil::Dispatch(*op_, {x, update->contiguous()}, outputs.get(), attrs));
       return outputs->at(0);
     } else {
       return OpInterpUtil::Dispatch<Tensor>(*op_, {x->contiguous(), update->contiguous()}, attrs);
